Struct rustfft::algorithm::DFT

pub struct DFT<T> { /* fields omitted */ }

Naive O(n^2 ) Discrete Fourier Transform implementation

This implementation is primarily used to test other FFT algorithms. In a few rare cases, such as small Cunningham Chain primes, this can be faster than the O(nlogn) algorithms

// Computes a naive DFT of size 1234
use rustfft::algorithm::DFT;
use rustfft::FFT;
use rustfft::num_complex::Complex;
use rustfft::num_traits::Zero;

let mut input:  Vec<Complex<f32>> = vec![Zero::zero(); 1234];
let mut output: Vec<Complex<f32>> = vec![Zero::zero(); 1234];

let dft = DFT::new(1234, false);
dft.process(&mut input, &mut output);

Methods

impl<T: FFTnum> DFT<T>

fn new(len: usize, inverse: bool) -> Self

Preallocates necessary arrays and precomputes necessary data to efficiently compute DFT