rustfft::algorithm

Struct GoodThomasAlgorithm

pub struct GoodThomasAlgorithm<T> { /* private fields */ }

Implementation of the Good-Thomas Algorithm (AKA Prime Factor Algorithm)

This algorithm factors a size n FFT into n1 * n2, where GCD(n1, n2) == 1

Conceptually, this algorithm is very similar to the Mixed-Radix, except because GCD(n1, n2) == 1 we can do some number theory trickery to reduce the number of floating-point multiplications and additions. Additionally, It can be faster than Mixed-Radix at sizes below 10,000 or so.

// Computes a forward FFT of size 1200, using the Good-Thomas Algorithm
use rustfft::algorithm::GoodThomasAlgorithm;
use rustfft::{Fft, FftPlanner};
use rustfft::num_complex::Complex;
use rustfft::num_traits::Zero;

let mut buffer = vec![Complex{ re: 0.0f32, im: 0.0f32 }; 1200];

// we need to find an n1 and n2 such that n1 * n2 == 1200 and GCD(n1, n2) == 1
// n1 = 48 and n2 = 25 satisfies this
let mut planner = FftPlanner::new();
let inner_fft_n1 = planner.plan_fft_forward(48);
let inner_fft_n2 = planner.plan_fft_forward(25);

// the good-thomas FFT length will be inner_fft_n1.len() * inner_fft_n2.len() = 1200
let fft = GoodThomasAlgorithm::new(inner_fft_n1, inner_fft_n2);
fft.process(&mut buffer);

Implementations

impl<T: FftNum> GoodThomasAlgorithm<T>

pub fn new(width_fft: Arc<dyn Fft<T>>, height_fft: Arc<dyn Fft<T>>) -> Self

Creates a FFT instance which will process inputs/outputs of size width_fft.len() * height_fft.len()

GCD(width_fft.len(), height_fft.len()) must be equal to 1