rustfft::algorithm

Struct RadersAlgorithm

pub struct RadersAlgorithm<T> {
    inner_fft: Arc<dyn Fft<T>>,
    inner_fft_data: Box<[Complex<T>]>,
    primitive_root: u64,
    primitive_root_inverse: u64,
    len: StrengthReducedU64,
    inplace_scratch_len: usize,
    outofplace_scratch_len: usize,
    immut_scratch_len: usize,
    direction: FftDirection,
}

Implementation of Rader’s Algorithm

This algorithm computes a prime-sized FFT in O(nlogn) time. It does this by converting this size-N FFT into a size-(N - 1) FFT, which is guaranteed to be composite.

The worst case for this algorithm is when (N - 1) is 2 * prime, resulting in a Cunningham Chain

// Computes a forward FFT of size 1201 (prime number), using Rader's Algorithm
use rustfft::algorithm::RadersAlgorithm;
use rustfft::{Fft, FftPlanner};
use rustfft::num_complex::Complex;

let mut buffer = vec![Complex{ re: 0.0f32, im: 0.0f32 }; 1201];

// plan a FFT of size n - 1 = 1200
let mut planner = FftPlanner::new();
let inner_fft = planner.plan_fft_forward(1200);

let fft = RadersAlgorithm::new(inner_fft);
fft.process(&mut buffer);

Rader’s Algorithm is relatively expensive compared to other FFT algorithms. Benchmarking shows that it is up to an order of magnitude slower than similar composite sizes. In the example size above of 1201, benchmarking shows that it takes 2.5x more time to compute than a FFT of size 1200.

Fields
inner_fft: Arc<dyn Fft<T>>inner_fft_data: Box<[Complex<T>]>primitive_root: u64primitive_root_inverse: u64len: StrengthReducedU64inplace_scratch_len: usizeoutofplace_scratch_len: usizeimmut_scratch_len: usizedirection: FftDirection
Implementations

impl<T: FftNum> RadersAlgorithm<T>

pub fn new(inner_fft: Arc<dyn Fft<T>>) -> Self

Creates a FFT instance which will process inputs/outputs of size inner_fft.len() + 1.

Note that this constructor is quite expensive to run; This algorithm must compute a FFT using inner_fft within the constructor. This further underlines the fact that Rader’s Algorithm is more expensive to run than other FFT algorithms

Panics

Panics if inner_fft.len() + 1 is not a prime number.

fn perform_fft_immut( &self, input: &[Complex<T>], output: &mut [Complex<T>], scratch: &mut [Complex<T>], )

fn perform_fft_out_of_place( &self, input: &mut [Complex<T>], output: &mut [Complex<T>], scratch: &mut [Complex<T>], )

fn perform_fft_inplace( &self, buffer: &mut [Complex<T>], scratch: &mut [Complex<T>], )