rustfft

Module array_utils

Structs

DoubleBuf 🔒

TransposeFactor 🔒

Traits

Load 🔒

LoadStore 🔒

Functions

bitreversed_transpose

compute_logarithm

factor_transpose 🔒

reverse_bits

reverse_remainders 🔒

transpose_small⚠

Given an array of size width * height, representing a flattened 2D array, transpose the rows and columns of that 2D array into the output benchmarking shows that loop tiling isn’t effective for small arrays (in the range of 50x50 or smaller)

validate_and_iter

validate_and_iter_unroll2x

validate_and_zip

validate_and_zip_mut

validate_and_zip_mut_unroll2x

validate_and_zip_unroll2x

workaround_transmute⚠

workaround_transmute_mut⚠