rustfft::avx

Module avx_vector

Macros

column_butterfly16_loadfn 🔒

column_butterfly32_loadfn 🔒

Structs

Rotation90

Traits

AvxArray

AvxArrayMut

AvxVector

A SIMD vector of complex numbers, stored with the real values and imaginary values interleaved. Implemented for __m128, __m128d, __m256, __m256d, but these all require the AVX instruction set.

AvxVector128

A 128-bit SIMD vector of complex numbers, stored with the real values and imaginary values interleaved. Implemented for __m128, __m128d, but these are all oriented around AVX, so don’t call methods on these from a SSE-only context

AvxVector256

A 256-bit SIMD vector of complex numbers, stored with the real values and imaginary values interleaved. Implemented for __m256, __m256d

Functions

apply_butterfly8_twiddle1⚠

apply_butterfly8_twiddle3⚠

mul_complex_conjugated 🔒 ⚠

Multiply the complex numbers in left by the complex numbers in right. This is exactly the same as mul_complex in AvxVector, but this implementation also conjugates the left input before multiplying

pairwise_complex_mul_assign_conjugated⚠

pairwise_complex_mul_conjugated⚠