rustfft::plan

Enum Recipe

pub(crate) enum Recipe {

    Dft(usize),
    MixedRadix {
        left_fft: Arc<Recipe>,
        right_fft: Arc<Recipe>,
    },
    GoodThomasAlgorithm {
        left_fft: Arc<Recipe>,
        right_fft: Arc<Recipe>,
    },
    MixedRadixSmall {
        left_fft: Arc<Recipe>,
        right_fft: Arc<Recipe>,
    },
    GoodThomasAlgorithmSmall {
        left_fft: Arc<Recipe>,
        right_fft: Arc<Recipe>,
    },
    RadersAlgorithm {
        inner_fft: Arc<Recipe>,
    },
    BluesteinsAlgorithm {
        len: usize,
        inner_fft: Arc<Recipe>,
    },
    RadixN {
        factors: Box<[RadixFactor]>,
        base_fft: Arc<Recipe>,
    },
    Radix4 {
        k: u32,
        base_fft: Arc<Recipe>,
    },
    Butterfly2,
    Butterfly3,
    Butterfly4,
    Butterfly5,
    Butterfly6,
    Butterfly7,
    Butterfly8,
    Butterfly9,
    Butterfly11,
    Butterfly12,
    Butterfly13,
    Butterfly16,
    Butterfly17,
    Butterfly19,
    Butterfly23,
    Butterfly24,
    Butterfly27,
    Butterfly29,
    Butterfly31,
    Butterfly32,

}

A Recipe is a structure that describes the design of a FFT, without actually creating it. It is used as a middle step in the planning process.

Variants

Dft(usize)

MixedRadix

Fields

left_fft: Arc<Recipe>

right_fft: Arc<Recipe>

GoodThomasAlgorithm

Fields

left_fft: Arc<Recipe>

right_fft: Arc<Recipe>

MixedRadixSmall

Fields

left_fft: Arc<Recipe>

right_fft: Arc<Recipe>

GoodThomasAlgorithmSmall

Fields

left_fft: Arc<Recipe>

right_fft: Arc<Recipe>

RadersAlgorithm

Fields

inner_fft: Arc<Recipe>

BluesteinsAlgorithm

Fields

len: usize

inner_fft: Arc<Recipe>

RadixN

Fields

factors: Box<[RadixFactor]>

base_fft: Arc<Recipe>

Radix4

Fields

k: u32

base_fft: Arc<Recipe>

Butterfly2

Butterfly3

Butterfly4

Butterfly5

Butterfly6

Butterfly7

Butterfly8

Butterfly9

Butterfly11

Butterfly12

Butterfly13

Butterfly16

Butterfly17

Butterfly19

Butterfly23

Butterfly24

Butterfly27

Butterfly29

Butterfly31

Butterfly32

Implementations

impl Recipe

pub fn len(&self) -> usize