rustfft::sse

Module sse_planner

Structs

FftPlannerSse

The SSE FFT planner creates new FFT algorithm instances using a mix of scalar and SSE accelerated algorithms. It requires at least SSE4.1, which is available on all reasonably recent x86_64 cpus.

Enums

Recipe

A Recipe is a structure that describes the design of a FFT, without actually creating it. It is used as a middle step in the planning process.

Constants

MAX_RADER_PRIME_FACTOR 🔒

MIN_RADIX4_BITS 🔒