Struct rustfft::FFT

pub struct FFT<T> {
    // some fields omitted
}

Methods

impl<T> FFT<T> where T: Signed + FromPrimitive + Copy

fn new(len: usize, inverse: bool) -> Self

Creates a new FFT context that will process signal of length len. If inverse is true, then this struct will run inverse FFTs. This implementation of the FFT doesn't do any scaling on both the forward and backward transforms, so doing a forward then backward FFT on a signal will scale the signal by its length.

fn process(&mut self, signal: &[Complex<T>], spectrum: &mut [Complex<T>])

Runs the FFT on the input signal buffer, and places the output in the spectrum buffer.

Panics

This method will panic if signal and spectrum are not the length specified in the struct's constructor.